data_storage/requirements.txt
kafka-python
confluent_kafka
mysql.connector

